WIP refactoring of the utils
