+++ http-utils/jvm/src/main/scala/io/iohk/atala/mediator/comm/MessageDispatcherJVM.scala
@@ -1,14 +1,14 @@
-package fmgp.did.comm
+package io.iohk.atala.mediator.comm
 
-import zio._
-import zio.json._
-import zio.http._
-import zio.http.model._
-
-import fmgp.did._
-import fmgp.did.comm._
-import fmgp.crypto.error._
-import fmgp.util.MyHeaders
+import fmgp.crypto.error.*
+import fmgp.did.*
+import fmgp.did.comm.*
+import io.iohk.atala.mediator.comm.*
+import io.iohk.atala.mediator.utils.MyHeaders
+import zio.*
+import zio.http.*
+import zio.http.model.*
+import zio.json.*
 
 object MessageDispatcherJVM {
   val layer: ZLayer[Client, Throwable, MessageDispatcher] =
