+++ http-utils/shared/src/main/scala/io/iohk/atala/mediator/comm/TapMessage.scala
@@ -1,7 +1,7 @@
-package fmgp.did.comm
+package io.iohk.atala.mediator.comm
 
-import zio.json._
-
+import zio.json.*
+import fmgp.did.comm.*
 final case class TapMessage(msg: EncryptedMessage, decrypted: PlaintextMessage)
 
 object TapMessage {
