+++ http-utils/shared/src/main/scala/io/iohk/atala/mediator/utils/MyHeaders.scala
@@ -1,4 +1,4 @@
-package fmgp.util
+package io.iohk.atala.mediator.utils
 
 object MyHeaders {
   final val xForwardedHost: CharSequence = "x-forwarded-host"
